battery-manager/requirements.txt
requests>=2.31.0
paho-mqtt>=2.0.0
python-dateutil>=2.9.0
pytz>=2024.1
